Refactor code to use Registry
I've implemented a Registry on my own just to learn how it works under the
hood. As this code is already committed, I'm changing the code to use the
official Registry making the tic tac toe code simpler.

Also, I'm refactoring the `GameSupervisor` renaming the functions to
tell more about the action than the internals.