Fix GameServer when there are no players in the game
When the last player left the game, the `GenServer` was returning the
error:

```
** (exit) exited in: GenServer.call({:via, Registry, {:game_server_registry, "teste"}}, {:leave, :o}, 5000)
    ** (EXIT) normal
    (elixir) lib/gen_server.ex:924: GenServer.call/3
```

That was happening because I changed the `leave` function to be a
`call`, but I forgot to change the stop returning.

When we want to stop a `GenServer`, if we are in a `handle_cast` we must
use the pattern:

```
{:stop, :normal, state}
```

Case we are in a `handle_call` we must must use:

```
{:stop, :normal, state, state}
```

Since the `handle_call` returns the GenServer state to the client.