Add name to game struct
Since every game has a name defined by the user, it makes sense to put
this name in the struct as well.

By now, this name is the same name I use to register and lookup after
with Registry, so it's unique and I think there is no problem in using
it.