Broadcast events in the GameEngine
From this commit, the game engine will start to broadcast two events:
:game_created and :game_closed. Then listener can react accordingly once
those events are dispatched.

To broadcast :game_closed, a refactor in the `leave` callback was
necessary. I just changed the tuple to use :closed instead of :ok once
the process stop, then I'm able to know if the process stopped.