Include assets installation in setup instructions
When I was setting the project up, I noticed that the command to install
the front-end dependencies was missing. I think it's good to include it
in the `README.md` especially because this project uses umbrella apps -
e.g. it was not clear for me that I had to go to the `game_ui` directory
in order to install them.