Move all dead views to Live View
This commit is a big refactor moving all dead views to Live View.

The main challenge here is that now we need to keep the User session in
the back-end since can't really do that with Live View since it works
over websockets.

I did that inspired by this blog post:

https://toranbillups.com/blog/archive/2020/06/26/cookie-authentication-with-phoenix-liveview/